refactor: rename DelimitedList to DelimitedCollection
Rename the `DelimitedList` mixin to `DelimitedCollection` (and the module
`_delimited_list.py` to `_delimited_collection.py`) ahead of generalizing it
beyond `list`. Also reorganize the converter tests to one file per converter:
the delimited-list tests move to `test_delimited_collection.py` and the
Counter tests to `test_counter_pivot_table.py`.

Pure rename and test reorganization; no behavior change.

Relates to #76.
