+++ fgmetric/converters/__init__.py
@@ -1,9 +1,9 @@
 from fgmetric.converters._counter_pivot_table import CounterPivotTable
-from fgmetric.converters._delimited_list import DelimitedList
+from fgmetric.converters._delimited_collection import DelimitedCollection
 from fgmetric.converters._null_sentinels import NullSentinels
 
 __all__ = [
     "CounterPivotTable",
-    "DelimitedList",
+    "DelimitedCollection",
     "NullSentinels",
 ]
